- Added detection of network registration timeout (15 minutes).* - Added hard reset for unresponsive module. - Stop other GSM services in case udev rules to not fire upon remove event.**
* HL8548's with firmware RHL85xx.5.5.18.0 from 2015/06/30 have been observed
to require 1 minute for "+CREG: 2,2" and are keeping 2 minutes "+CREG: 2,3"
until changing to "+CREG: 2,1" and then failing on dial in on their very
first use (factory new after soldering process).
A module reset during the "+CREG: 2,3" is not able the shorten that
initial behaviour.
A timeout of 15 minutes should not interfere with this lengthy process.

**The cause for udev rules not being fired is currently unclear.