ppp0: don't set the dns configuration in ppp0.network
The current version of systemd-networkd doesn't support the
KeepConfiguration option. Therefore, it might happen that the dns
configuration in ppp0.netowrk is set after the interface got it's ip
address through dhcp. This will then remove the ip addresses because
this option is missing. It seems this happens more often when we
optimize the boot time but could also have happened in the past.