Carry out communication module startup depending on type (HL85xx or HL78xx).
With the previous startup procedure, the GSM_ON_N for the HL78xx was not
handled as required in its technical specification on purpose to not have
to handle a type dependent startup procedure.
However, at least the UART interface of the HL78xx seems to have an issue
with the violation of the GSM_ON_N's requirements.
Considering the previous issues with the HL78xx's startup procedure, a type
dependent startup procedure is considered a necessity.