mnt-log: do lazy umount when stopping
Do a lazy umount when stopping the mnt-log task. With this we get rid of
a dependency issue on shutdown. journald is still having a file open
under /var/log/journal. Therefore, we can't umount. However, we mount
late in the boot stage (because it is slow) and therefore we need to do
a lazy mount (the umount will happen once all references are destroyed).