Replaced update file HL7802_4.3.9.0_4.4.14.0 by HL7802_4.4.14.0_to_4.6.9.4.
All OMNI LTE-M devices were produced and released in the field with 4.4.14.0.
Either they had the correct SKU 1104569 (USB, 4.4.14.0) in production or were
in case of 750 devices "reworked" from SKU 1104503 (UART, 4.3.9.0) to the
desired one.

Since SKU 1104569 (USB, 4.4.14.0) is going to be replaced by SKU 1105030
(USB, 4.6.9.4), it makes sense to not only upgrade to the first version
supporting AT+KSELACQ "4.5.4.0" but directly "4.6.9.4".
This way, there will be two rather than three different HL7802 firmware
versions out in the field.