Stop medusa-Dataserver.service in a loop in case the stop job is cancelled.
In some cases, it could be observed that systemd cancelled stopping the
medusa-Dataserver.service when the USB drive containing a firmware update
was plugged in already at startup.
The cause of the cancellation is unknown but assumed to be related to either
high CPU load and/or timing.
Stopping medusa-Dataserver.service in a loop until success is more robust
and should always stop the DataServer based applications to have all possible
CPU resources available for a faster installation of the firmware.