Increased RuntimeWatchdogSec value to 1min.
Since especially now, with having even more medusa applications running as RT,
a call to "systemctl daemon-reload" takes a couple of seconds and the
RuntimeWatchdog is not always triggered within time (5s as half of the value
in worst case).

Although "systemctl daemon-reload" is only used for development, there might
be other commands/situations where this could occur and a higher value
is the safer way to go.