Detect invalid/cutoff line in .candump and add it (incl. subsequent lines) as comment in .trc file.
The reason (for converting) and adding all subsequent lines as PEAK comment
(";") is to not end up with a .trc file where some messages are missing when
inspecting it in PCAN Explorer.

As of now, it is unclear as to why an invalid/cutoff line is created in a .candump.
The best guess is that it might be a timing/buffering issue that is now showing
up more frequently than in the past.

The case when the first line is invalid, resp. likely the remaining part
after the cutoff, is not handled.