Combined candump & candump-manual into one service to lower CPU load.
The main reason for adding candump-manual was to have a candump with
bigger size (initially 2x16MB) than the normal one (2x4MB) and whose
buffer is also not rotated on EnergyBusController Warnings/Errors and
thus shortening it.
With the usage of LZOP and the increase of chunks (64x4MB uncompressed),
it should now be acceptable to use the normal buffer for both manual and
Warnings/Error triggers.