Removed cpupower service (keeping started.target for explanatory purpose).
Did not have an impact due to CONFIG_CPU_FREQ_DEFAULT_GOV_ONDEMAND=y in defconfig.

Concerning the governors, there does not seem to be a tangible benefit of using
the "ondemand" over "performance" one, as the CPU voltage is kept to 1250000uV
for both the 6UL (G2/528MHz) and 6ULL (Y2/792MHz) variants being used.
Ideally, the usage of "performance" might help with reducing the latency
for especially RT scheduled tasks.