Use nc instead of nslookup to determine connection to DNS servers.
The nslookup from busybox has no option to set a timeout value and
especially if a firmware update is being downloaded and the device
is connected via NB-IoT (or also in case of bad reception), timeouts
are easily observable.

Since adding bind-utils (for nslookup or dig) seems overkill and an
open port check will suffice to check the connectivity, there should
be no drawback of using nc.