Added hosts-static.service to facilitate pre-fetching IP of backend (#9201).
It is imperative that /etc/hosts, resp. the "localhost" is always available
or at least as early as possible and not just a symlink to a file that is
populated later on.

There are tools (such as gpsd) which rely on localhost being resolvable and
misbehave (stay in a non-working state while hogging the CPU indefinitely).

The /etc/hosts file as created during the build is thus installed as
/etc/hosts.static, a symlink to /tmp/hosts created and copied to the
where the symlink points within the scope of local-fs.target.