Use default instruction set (thumb) defined by Yocto.
ARM_INSTRUCTION_SET was set/kept to "arm" within the scope of a sumo (2.5) to
dunfell (3.1) update to keep the patch files for the update smaller.
However, with kirkstone (4.0) and beyond, too many things have changed and
keeping the instruction set does no longer provide a benefit concerning size.
On the contrary, patch files from sumo based releases to the current master
are actually smaller, when thumb is used as an instruction set.
Furthermore, the effective space used on the ubifs partition is actually
smaller when using thumb (about 55%) than arm (about 66%) because the
underlying ubifs compression method can handle it better. Even if compressing
a thumb and arm tarball of the image with xz indicates otherwise.