Added new service systemd-timesyncd-prepare.service for /mnt/data/timesync/.
With the current version of systemd, the directory is not created by
systemd-timesyncd itself but it relies on the service unit creating
the state directory (via StateDirectory=systemd/timesync).
Since /mnt/data/timesync shall be used as the state directory (as changed by
0001-move-clock-state-file-to-data-partition.patch), the creation of the
directory has to be manually handled.
This is especially relevant when not coming from an older firmware/systemd
where this directory has been already created (or after a purgedata);
A missing /mnt/data/timesync will prevent the systemd-timesyncd service from
starting.